requirements.txt
torch
torchvision
pillow
numpy
matplotlib
pandas
transformers
streamlit
ultralytics